requirements.txt
pytest
python-docx
reportlab
PyPDF2
requests
